make linter happy, initial update for internal references to Makers
