+++ gnrpy/gnr/app/gnrdeploy.py
@@ -360,735 +360,6 @@ def projectBag(project_name,packages=None,branches=None,exclude_branches=None):
     return result
 
     
-class ProjectMaker(object):
-    """Handle the autocreation of a package.
-    
-    To autocreate a package, please type in your console::
-        
-        gnrmkpackage packagename
-        
-    where ``packagename`` is the name of your ``package`` folder."""
-    def __init__(self, project_name, base_path=None):
-        self.project_name = project_name
-        self.base_path = base_path or '.'
-        
-    def do(self):
-        """Create the project path with its subfolders paths: the path of the ``packages`` folder,
-        the path of the ``sites`` folder, the path of the ``instances`` folder and the path of the
-        ``resources`` folder."""
-        self.project_path = os.path.join(self.base_path, self.project_name)
-        self.packages_path = os.path.join(self.project_path, 'packages')
-        self.instances_path = os.path.join(self.project_path, 'instances')
-        for path in (self.project_path, self.packages_path, self.instances_path):
-            if not os.path.isdir(path):
-                os.makedirs(path, exist_ok=True)
-                
-
-
-class SiteMaker(object):
-    #deprecated 
-    """Handle the autocreation of the ``sites`` folder.
-    
-    To autocreate the ``sites`` folder, please type in your console::
-        
-        gnrmksite sitesname
-        
-    where 'sitesname' is the name of your ``sites`` folder.
-    """
-    def __init__(self, site_name, base_path=None, resources=None, instance=None, dojo_version='11',
-                 wsgi_port=None, wsgi_reload=None, wsgi_mainpackage=None, wsgi_debug=None, config=None):
-        self.site_name = site_name
-        self.base_path = base_path or '.'
-        self.resources = resources or []
-        self.instance = instance
-        self.wsgi_port = wsgi_port
-        self.wsgi_reload = wsgi_reload
-        self.wsgi_mainpackage = wsgi_mainpackage
-        self.wsgi_debug = wsgi_debug
-        self.dojo_version = dojo_version
-        self.config = config
-        
-    def do(self):
-        """TODO"""
-        self.site_path = os.path.join(self.base_path, self.site_name)
-        pages_path = os.path.join(self.site_path, 'pages')
-        root_py_path = os.path.join(self.site_path, 'root.py')
-        siteconfig_xml_path = os.path.join(self.site_path, 'siteconfig.xml')
-        if not os.path.isdir(self.site_path):
-            os.makedirs(self.site_path, exist_ok=True)
-        if not os.path.isdir(pages_path):
-            os.makedirs(pages_path, exist_ok=True)
-        if not os.path.isfile(root_py_path):
-            root_py = open(root_py_path, 'w')
-            root_py.write("""#!/usr/bin/env python2.6
-import sys
-sys.stdout = sys.stderr
-from gnr.web.gnrwsgisite import GnrWsgiSite
-site = GnrWsgiSite(__file__)
-
-def application(environ,start_response):
-    return site(environ,start_response)
-
-if __name__ == '__main__':
-    from gnr.web.server import NewServer
-    server=NewServer(__file__)
-    server.run()""")
-            root_py.close()
-        if not os.path.isfile(siteconfig_xml_path):
-            if not self.config:
-                siteconfig = Bag()
-                if self.instance:
-                    siteconfig.setItem('instances.%s' % self.instance, None)
-                for resource in self.resources:
-                    if isinstance(resource, tuple) or isinstance(resource, list):
-                        resource, resource_path = resource
-                        siteconfig.setItem('resources.%s' % resource, None, path=resource_path)
-                    else:
-                        siteconfig.setItem('resources.%s' % resource, None)
-                wsgi_options = dict()
-                for option in ('reload', 'debug', 'port', 'mainpackage'):
-                    value = getattr(self, 'wsgi_%s' % option, None)
-                    if value:
-                        wsgi_options[option] = value
-                siteconfig.setItem('wsgi', None, **wsgi_options)
-                siteconfig['connection_timeout'] = None
-                siteconfig['connection_refresh'] = None
-                siteconfig.setItem('dojo', None, version=self.dojo_version)
-            else:
-                siteconfig = self.config
-            siteconfig.toXml(siteconfig_xml_path,typevalue=False,pretty=True)
-            
-class InstanceMaker(object):
-    """Handle the autocreation of the ``instances`` folder.
-    
-    To autocreate the ``instances`` folder, please type in your console::
-        
-        gnrmkinstance instancesname
-        
-    where ``instancesname`` is the name of your ``instances`` folder.
-    """
-    def __init__(self, instance_name, base_path=None, packages=None, authentication=True, authentication_pkg=None,
-                 db_dbname=None, db_implementation=None, db_host=None, db_port=None,
-                 db_user=None, db_password=None, use_dbstores=False, config=None,main_package=None):
-        self.instance_name = instance_name
-        self.base_path = base_path or '.'
-        self.packages = packages or []
-        self.db_dbname = db_dbname or instance_name
-        self.authentication = authentication
-        self.main_package = main_package
-        if self.authentication:
-            self.authentication_pkg = authentication_pkg
-            if not self.authentication_pkg and self.packages:
-                package = self.packages[0]
-                if isinstance(package, tuple) or isinstance(package, list):
-                    package = package[0]
-                self.authentication_pkg = package
-            if not self.authentication_pkg:
-                self.authentication_pkg = 'adm'
-        self.db_implementation = db_implementation
-        self.db_host = db_host
-        self.db_port = db_port
-        self.db_user = db_user
-        self.db_password = db_password
-        self.use_dbstores = use_dbstores
-        self.config = config
-        self.instance_path = os.path.join(self.base_path, self.instance_name)
-        self.config_path = os.path.join(self.instance_path,'config')
-    
-    def do(self):
-        instanceconfig = self.do_instance()
-        self.do_site(instanceconfig)
-        instanceconfig_xml_path = os.path.join(self.config_path, 'instanceconfig.xml')
-        instanceconfig.toXml(instanceconfig_xml_path,typevalue=False,pretty=True)
-
-    def do_instance(self):
-        custom_path = os.path.join(self.instance_path, 'custom')
-        data_path = os.path.join(self.instance_path, 'data')
-        instanceconfig_xml_path = os.path.join(self.config_path, 'instanceconfig.xml')
-        folders_to_make = [self.instance_path,self.config_path, custom_path, data_path]
-        if self.use_dbstores:
-            dbstores_path = os.path.join(self.instance_path, 'dbstores')
-            folders_to_make.append(dbstores_path)
-        for path in folders_to_make:
-            if not os.path.isdir(path):
-                os.makedirs(path, exist_ok=True)
-        if not os.path.isfile(instanceconfig_xml_path):
-            if not self.config:
-                instanceconfig = Bag()
-                db_options = dict()
-                for option in ('dbname', 'implementation', 'host', 'port', 'username', 'password'):
-                    value = getattr(self, 'db_%s' % option, None)
-                    if value:
-                        db_options[option] = value
-                instanceconfig.setItem('db', None, **db_options)
-                instanceconfig.setItem('packages', None)
-                for package in self.packages:
-                    if isinstance(package, tuple) or isinstance(package, list):
-                        package, package_path = package
-                        instanceconfig.setItem('packages.%s' % package.replace(':','_'), None, path=package_path,pkgcode=package)
-                    else:
-                        instanceconfig.setItem('packages.%s' % package.replace(':','_'), None,pkgcode=package)
-                if self.authentication:
-                    instanceconfig.setItem('authentication', None, pkg=self.authentication_pkg)
-                    instanceconfig.setItem('authentication.py_auth', None, defaultTags="user", pkg="adm",
-                                           method="authenticate")
-            else:
-                instanceconfig = self.config
-            return instanceconfig
-
-    def do_site(self,instanceconfig):
-        """TODO"""
-        self.site_path = os.path.join(self.instance_path, 'site')
-        root_py_path = os.path.join(self.instance_path, 'root.py')
-        siteconfig_xml_path = os.path.join(self.config_path, 'siteconfig.xml')
-        if not os.path.isdir(self.site_path):
-            os.makedirs(self.site_path, exist_ok=True)
-        if not os.path.isfile(root_py_path):
-            root_py = open(root_py_path, 'w')
-            root_py.write("""
-import sys
-sys.stdout = sys.stderr
-from gnr.web.gnrwsgisite import GnrWsgiSite
-site = GnrWsgiSite(__file__)
-
-def application(environ,start_response):
-    return site(environ,start_response)
-
-if __name__ == '__main__':
-    from gnr.web.server import NewServer
-    server=NewServer(__file__)
-    server.run()""")
-            root_py.close()
-        if not os.path.isfile(siteconfig_xml_path):
-            if not self.config:
-                siteconfig = Bag()
-                wsgi_options = dict()
-                wsgi_options.setdefault('mainpackage',self.main_package)
-                for option in ('reload', 'debug', 'port', 'mainpackage'):
-                    value = getattr(self, 'wsgi_%s' % option, None)
-                    if value:
-                        wsgi_options[option] = value
-                siteconfig.setItem('wsgi', None, **wsgi_options)
-            else:
-                siteconfig = self.config
-            instanceconfig.addItem('site',siteconfig)
-
-            
-class PackageMaker(object):
-    """Handle the autocreation of the ``packages`` folder.
-    
-    To autocreate the ``packages`` folder, please type in your console::
-        
-        gnr dev mkpackage packagesname
-        
-    where ``packagesname`` is the name of your ``packages`` folder.
-    """
-    def __init__(self, package_name, base_path=None, sqlschema=None, sqlprefix=True,
-                 name_short=None, name_long=None, name_full=None,
-                 login_url=None, comment=None,helloworld=False):
-        self.package_name = package_name
-        self.base_path = base_path or '.'
-        self.name_short = name_short or self.package_name.capitalize()
-        self.name_full = name_full or self.package_name.capitalize()
-        self.name_long = name_long or self.package_name.capitalize()
-        self.sqlschema = sqlschema or self.package_name.lower()
-        self.sqlprefix = sqlprefix
-        self.comment = comment or '%s package' % self.package_name
-        self.login_url = login_url or '%s/login' % self.package_name
-        self.helloworld = helloworld
-        self.package_path = os.path.join(self.base_path, self.package_name)
-        self.model_path = os.path.join(self.package_path, 'model')
-        self.cli_path = os.path.join(self.package_path, 'cli')
-        self.lib_path = os.path.join(self.package_path, 'lib')
-        self.webpages_path = os.path.join(self.package_path, 'webpages')
-        self.resources_path = os.path.join(self.package_path, 'resources')
-        self.framedindex_path = os.path.join(self.webpages_path,'index.py')
-        self.main_py_path = os.path.join(self.package_path, 'main.py')
-        
-    def do(self):
-        """Creates the files of the ``packages`` folder"""
-        for path in (self.package_path, self.model_path, self.cli_path,
-                     self.lib_path, self.webpages_path, self.resources_path):
-            if not os.path.isdir(path):
-                os.makedirs(path, exist_ok=True)
-
-        # create an emptydir file allowing an empty cli directory to be
-        # pushed to repository
-        open(os.path.join(self.cli_path, ".emptydir"), "w").close()
-        
-        # create an empty requirements.txt file, hopefully developers
-        # will be reminded by its presence that dependencies can be added
-        # in this file
-        open(os.path.join(self.package_path, 'requirements.txt'), "w").close()
-
-        # create an placeholder README.md file, since it's a policy
-        # that has been established to have a README for each package
-        # refs #83
-        with open(os.path.join(self.package_path, 'README.md'), "w") as readme_fp:
-            head = f"Package '{self.package_name}'"
-            readme_fp.write(head)
-            readme_fp.write("\n")
-            readme_fp.write("="*len(head))
-            
-            
-        sqlprefixstring = ''
-        if not os.path.exists(self.main_py_path):
-            if self.sqlprefix is not None:
-                if self.sqlprefix not in (True, False):
-                    self.sqlprefix = "'%s'" % self.sqlprefix
-                sqlprefixstring = "sqlprefix=%s,"%(self.sqlprefix)
-            main_py_options = dict(comment=self.comment, sqlschema=self.sqlschema, sqlprefixstring=sqlprefixstring, name_short=self.name_short,
-                                   name_long=self.name_long, name_full=self.name_full, login_url=self.login_url)
-            main_py = open(self.main_py_path, 'w')
-            main_py.write("""#!/usr/bin/env python
-# encoding: utf-8
-from gnr.app.gnrdbo import GnrDboTable, GnrDboPackage
-
-class Package(GnrDboPackage):
-    def config_attributes(self):
-        return dict(comment='%(comment)s',sqlschema='%(sqlschema)s',%(sqlprefixstring)s
-                    name_short='%(name_short)s', name_long='%(name_long)s', name_full='%(name_full)s')
-                    
-    def config_db(self, pkg):
-        pass
-        
-class Table(GnrDboTable):
-    pass
-""" % main_py_options)
-            main_py.close()
-
-        if not os.path.exists(self.framedindex_path):
-            with open(self.framedindex_path, 'w') as indexpage:
-                indexpage.write("""# -*- coding: utf-8 -*-
-            
-class GnrCustomWebPage(object):
-    py_requires = 'plainindex'
-    """)
-        if self.helloworld:
-            with open(os.path.join(self.webpages_path,'hello_world.py'), 'w') as helloworld:
-                helloworld.write("""# -*- coding: utf-8 -*-
-            
-class GnrCustomWebPage(object):
-    def main_root(self,root,**kwargs):
-        root.h1('Hello world',text_align='center')
-    """)
-            
-class ResourceMaker(object):
-    """Handle the creation of the ``resources`` folder"""
-    def __init__(self, resource_name, base_path=None):
-        self.resource_name = resource_name
-        self.base_path = base_path or '.'
-        
-    def do(self):
-        """TODO"""
-        self.resource_path = os.path.join(self.base_path, self.resource_name)
-        for path in (self.resource_path, ):
-            if not os.path.isdir(path):
-                os.makedirs(path, exist_ok=True)
-        
-class ThPackageResourceMaker(object):
-    def __init__(self, application, package=None, tables=None, force=False,
-                 menu=False, columns=2, guess_size=False, indent=4,
-                 bag_columns=None,
-                 filename=None, output=None):
-        self.option_force = force
-        self.option_menu = menu
-        self.option_output = filename or output
-        self.option_columns = columns
-        self.option_guess_size = guess_size
-        if self.option_guess_size:
-            logger.debug("Guessing column width by data size: ACTIVE")
-        self.option_indent = indent
-        self.pkg_tables = defaultdict(list)
-        self.pkg_menus = dict()
-        self.app = application 
-        self.package = package
-        self.bag_columns=bag_columns or dict(view=False, form=False)
-        self.tables = tables if tables else list(self.app.db.packages[self.package].tables.keys())
-        if not self.app.packages(package):
-            raise ModuleNotFoundError(f"Package {package} was not found")
-        
-        self.packageFolder = self.app.packages(package).packageFolder
-        self.has_lookups = False
-        
-    def makeResources(self):
-        for table in self.tables:
-            table_full_name = f"{self.package}.{table}"
-            logger.debug("Processing table %s", table_full_name)
-            if 'lookup' in self.app.db.table(table_full_name).attributes:
-                logger.debug("Skipping lookup table %s", table_full_name)
-                self.has_lookups = True
-                continue
-            self.createResourceFile(table)
-
-        if self.option_menu:
-            self.makeMenu()
-
-    def makeMenu(self):
-        menu_xml_path = os.path.join(self.packageFolder, 'menu.xml')
-        self.pkg_menus[self.package] = Bag(menu_xml_path) if os.path.exists(menu_xml_path) else Bag()
-
-        if self.has_lookups:
-            self.pkg_menus[self.package].setItem('auto.lookups', None,
-                                                 label='!!Lookup tables',
-                                                 pkg=self.package, tag='lookupBranch')
-                                                 
-        menupath = Path(self.packageFolder) / 'menu.py'
-        
-        if menupath.exists():
-            logger.info("Menu file %s exists, appending", menupath)
-            source = menupath.read_text(encoding="utf-8")
-            tree = ast.parse(source)
-            lines = source.splitlines(keepends=True)
-            class Finder(ast.NodeVisitor):
-                """
-                Simple AST visitor to search for the root branch assignment
-                and add a new page to such branch
-                """
-                def __init__(self):
-                    self.config_func = None
-                    self.branch_var = None
-                def visit_ClassDef(self, node):
-                    if node.name != "Menu":
-                        return
-                    for item in node.body:
-                        if isinstance(item, ast.FunctionDef) and item.name == "config":
-                            self.config_func = item
-                            self._find_branch_assignment(item)
-                            return
-                def _find_branch_assignment(self, func):
-                    for stmt in func.body:
-                        if not isinstance(stmt, ast.Assign):
-                            continue
-                        if len(stmt.targets) != 1:
-                            continue
-                        target = stmt.targets[0]
-                        if not isinstance(target, ast.Name):
-                            continue
-                        value = stmt.value
-                        if not isinstance(value, ast.Call):
-                            continue
-                        func_attr = value.func
-                        if (
-                                isinstance(func_attr, ast.Attribute)
-                                and isinstance(func_attr.value, ast.Name)
-                                and func_attr.value.id == "root"
-                                and func_attr.attr == "branch"
-                        ):
-                            self.branch_var = target.id
-                            return
-
-            finder = Finder()
-            finder.visit(tree)
-            if not finder.config_func:
-                raise RuntimeError("Menu.config not found")
-            if finder.branch_var is None:
-                raise RuntimeError("Root branch not found")
-            insert_at = finder.config_func.end_lineno
-            for item in self.pkg_menus[self.package]['auto'] or []:
-                new_line = f"\n        {finder.branch_var}.thpage(u'{item.attr['label']}', table='{item.attr['table']}')"
-                lines.insert(insert_at, new_line)
-                insert_at += 1
-
-            menupath.write_text("".join(lines), encoding="utf-8")
-
-        else:
-            logger.info("Menu file %s does not exist, creating", menupath)
-            m = MenuStruct()
-            for t in self.tables:
-                tblobj = self.app.db.table('%s.%s' %(self.package,t))
-                if tblobj.attributes.get('lookup'):
-                    self.has_lookups = True
-                else:
-                    m.thpage(tblobj.name_plural or tblobj.name_long, table=tblobj.fullname)
-            if self.has_lookups:
-                m.lookupBranch("Lookup tables", pkg=self.package)
-            m.toPython(menupath)
-        if os.path.exists(menu_xml_path):
-            os.remove(menu_xml_path)
-    def write(self,line=None, indent=0):
-        line = line or ''
-        self.out_file.write('%s%s\n'%(self.option_indent*indent*' ',line))
-
-    def writeHeaders(self):
-        self.write('#!/usr/bin/python3')
-        self.write('# -*- coding: utf-8 -*-')
-        self.write()
-    
-    def writeImports(self):
-        self.write("from gnr.web.gnrbaseclasses import BaseComponent")
-        self.write()
-    
-    def writeViewClass(self, tblobj, column_groups):
-        self.write("class View(BaseComponent):")
-        self.write()
-        self.write("def th_struct(self,struct):", indent=1)
-        self.write('r = struct.view().rows()', indent=2)
-        for group, columns in column_groups.items():
-            if group:
-                self.write(f"{group[0]}_col_group = r.columnset('{group[0]}', name='{group[1]}')", indent = 2)
-                for column, size in columns:
-                    if self.option_guess_size:
-                        self.write("%s_col_group.fieldcell('%s', width='%iem')" % (group[0], column, size), indent=2)
-                    else:
-                        self.write("%s_col_group.fieldcell('%s')" % (group[0], column), indent=2)
-            else:
-                for column, size in columns:
-                    if self.option_guess_size:
-                        self.write("r.fieldcell('%s', width='%iem')"%(column,size), indent=2)
-                    else:
-                        self.write("r.fieldcell('%s')"%column, indent=2)
-        self.write()
-        self.write("def th_order(self):", indent=1)
-        self.write("return '%s'" % columns[0][0], indent=2)
-        self.write()
-        self.write("def th_query(self):", indent=1)
-        searchcol = tblobj.attributes.get('caption_field') 
-        
-        if not searchcol:
-            l = [c for c in list(tblobj.columns.values()) if c.attributes.get('dtype') in ('A','T','C') and not c.attributes.get('_sysfield')]
-            if l:
-                searchcol = l[0].name
-        self.write("return dict(column='%s', op='contains', val='')"%searchcol, indent=2)
-        self.write()
-        self.write()
-        self.write()
-
-    def writeFormClass(self, tblobj, columns):
-        children = self.getChildrenRelations(tblobj)
-        self.write("class Form(BaseComponent):")
-        self.write()
-        self.write("def th_form(self, form):", indent=1)
-        if children:
-            self.write("bc = form.center.borderContainer()", indent=2)
-            self.write("top = bc.contentPane(region='top',datapath='.record')", indent=2)
-            self.write("fb = top.formbuilder(cols=%i, border_spacing='4px')"%self.option_columns, indent=2)
-            for column, size, dtype in columns:
-                self.write("fb.field('%s')"%column, indent=2)
-            if len(children)>1:
-                self.write("tc = bc.tabContainer(region='center',margin='2px')", indent=2)
-                for c,mode in children:
-                    self.write("tab_%s = tc.contentPane(title='%s')" %(c.replace('@',''),(tblobj.name_plural or tblobj.name_long)), indent=2)
-                    self.write("tab_%s.%sTableHandler(relation='%s')" %(c.replace('@',''),mode,c), indent=2)
-            else:
-                c,mode = children[0]
-                self.write("center = bc.contentPane(region='center')", indent=2)
-                self.write("center.%sTableHandler(relation='%s')" %(mode,c), indent=2)
-        else:
-            self.write("pane = form.record", indent=2)
-            self.write("fb = pane.formbuilder(cols=%i, border_spacing='4px')"%self.option_columns, indent=2)
-            for column, size, dtype in columns:
-                tag=''
-                if dtype=='X':
-                    if isinstance( self.bag_columns['form'], str):
-                        tag = ", tag='%s'" %  self.bag_columns['form'] 
-                self.write("fb.field('%s' %s)"% (column, tag), indent=2)
-                
-        self.write()
-        self.write()
-        self.write("def th_options(self):", indent=1)
-        hierarchical  = tblobj.column('hierarchical_pkey') is not None
-        hierarchical_chunk = ''
-        if hierarchical:
-            hierarchical_chunk = ', hierarchical=True'
-        self.write("return dict(dialog_height='400px', dialog_width='600px' %s)" %hierarchical_chunk, indent=2)
-
-
-    def getChildrenRelations(self,tblobj):
-        result = []
-        for relation,j in tblobj.relations.digest('#k,#a.joiner'):
-            if j and j['mode'] == 'M' and j.get('thmode'):
-                result.append((relation,j.get('thmode')))
-        return result
-
-    def columnWidthEstimate(self, column):
-        """
-        Estimates the width of a column based on its data type (dtype) and specified size attributes.
-        
-        Description
-        -----------
-        For text columns ('A' or 'T'), it uses a conversion map
-        to determine the appropriate width. For date ('D') and datetime ('DH') columns, it returns
-        fixed widths. If the data type is not recognized, it returns a default width.
-
-        Parameters
-        ----------
-        column : Column
-            The column object containing attributes such as dtype and size.
-
-        Returns
-        -------
-        int
-            The calculated width of the column. 
-        """
-        # TODO: Review the LUT and evaluate the use of a formula.
-        # Note that the commented formulae seems to be less accurate than the LUT
-
-        MIN_WIDTH = 2
-        MAX_SIZE = 50
-        DEFAULT_WIDTH = 7
-
-        sizeWidthMap = {
-            1: 2,
-            2: 2,
-            3: 3,
-            4: 4,
-            5: 4,
-            6: 5,
-            7: 5,
-            8: 6,
-            9: 6,
-            10: 6,
-            11: 6,
-            12: 7,
-            13: 7,
-            14: 8,
-            15: 8,
-            16: 9,
-            17: 9,
-            18: 10,
-            19: 10,
-            20: 10,
-            21: 11,
-            22: 11,
-            23: 11,
-            24: 12,
-            25: 12,
-            26: 12,
-            27: 13,
-            28: 13,
-            29: 14,
-            30: 14,
-            31: 15,
-            32: 15,
-            33: 15,
-            34: 16,
-            35: 17,
-            36: 17,
-            37: 18,
-            38: 18,
-            39: 18,
-            40: 19,
-            41: 19,
-            42: 20,
-            43: 20,
-            44: 21,
-            45: 21,
-            46: 22,
-            47: 22,
-            48: 23,
-            49: 23,
-            50: 24
-        }
-
-        def handleTextColumn(column):
-            try:
-                sizeTxt = column.attributes.get('size', '')
-            except Exception as e:
-                logger.warning(f'Unable to get size attribute for column {column.name}: {e}. Using default width.')
-                return DEFAULT_WIDTH
-            
-            if not sizeTxt:
-                logger.warning(f'No size attribute found for column {column.name}. Using default width.')
-                return DEFAULT_WIDTH
-
-            if ':' in sizeTxt:
-                size = int(sizeTxt.split(':')[1])
-            else:
-                size = int(sizeTxt)
-
-            if size > max(sizeWidthMap):
-                logger.info(f'Column {column.name} has a size greater than the maximum. Topping to max handled value.')
-                return sizeWidthMap[MAX_SIZE]
-            # LUT conversion
-            return sizeWidthMap[size]
-            
-            # ChatGPT 4o conversion
-            # return round(0.429 * size + 1.709)
-
-            # ChatGPT o3-mini-high conversion
-            # return round(0.45 * size + 1.55)
-
-            # Deepseek conversion
-            # return round(0.44 * size + 1.56)
-
-            # Claude conversion
-            # return floor(2.8 * size^0.45)
-
-        typesHandler = {
-            'A': handleTextColumn,  # Varchar
-            'T': handleTextColumn,  # Text
-            'C': handleTextColumn,  # Chars
-            'D': 6,       # Date
-            'H': 4,       # Time
-            'DH': 9,      # DateTime
-            'DHZ': 12,    # DateTime + TimeZone
-            'B': 3,       # Boolean
-            'N': 7,       # Numeric
-            'L': 7,       # Numeric (long)
-            'I': 7,       # Numeric (int)
-            'R': 7,       # Numeric (float)
-            'X': 10,      # Bag
-        }
-        handler = typesHandler.get(column.dtype, 7)
-        if callable(handler):
-            return handler(column)
-        return handler
-
-    def createResourceFile(self, table):
-        resourceFolder = os.path.join(self.packageFolder,'resources', 'tables', table)
-        if not os.path.exists(resourceFolder):
-            os.makedirs(resourceFolder, exist_ok=True)
-        name = 'th_%s.py'%table
-        path = os.path.join(resourceFolder, name) if not self.option_output else self.option_output
-        if os.path.exists(path) and not self.option_force:
-            logger.warning('%s exist: will be skipped, use -f/--force to force replace', name)
-            return
-        
-        column_groups = defaultdict(list)
-        columns = []
-        tbl_obj =  self.app.db.table('%s.%s'%(self.package,table))
-
-        for col_name, column in tbl_obj.columns.items():
-
-            # Skip id and internal columns
-            if col_name=='id' or column.attributes.get('_sysfield') or col_name.startswith('__') or column.type == 'O':
-                continue
-
-            logger.debug(f'Processing column {column.name}')
-            if self.option_guess_size:
-                width = self.columnWidthEstimate(column)
-                logger.debug(f'Estimated width for column {column.name}: {width}em')
-            else:
-                width = 7
-
-            col_group_label = column.attributes.get("colgroup_label", None)
-
-            if col_group_label is None:
-                col_group = ""
-            else:
-                col_group = (
-                    col_group_label,
-                    column.attributes.get("colgroup_name_long")
-                )
-
-            column_groups[col_group].append((column.name,width))
-            columns.append((column.name, width))
-
-        if not columns:
-            logger.error("Table %s does not contain any valid column", table)
-            return
-
-        try:
-            with open(path,'w') as out_file:
-                self.out_file = out_file
-                self.writeHeaders()
-                self.writeImports()
-                self.writeViewClass(tbl_obj, column_groups)
-                self.writeFormClass(tbl_obj, columns)
-                logger.info('%s created', name)
-        except Exception as e:
-            logger.exception(f'Error creating output file: {path}. Error: {e}')
 
 ################################# DEPLOY CONF BUILDERS ################################
