+++ gnrpy/gnr/dev/cli/gnrmkpackage.py
@@ -4,8 +4,8 @@
 import os
 
 from gnr.core.cli import GnrCliArgParse
-from gnr.app.gnrdeploy import PackageMaker
 from gnr.app.pathresolver import PathResolver
+from gnr.dev.makers.package import PackageMaker
 
 description = """Create a new Genropy package"""
